\name{bceQTL-package}
\alias{bceQTL-package}
\alias{bceQTL}
\docType{package}
\title{
\packageTitle{bceQTL}
}
\description{
\packageDescription{bceQTL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bceQTL}
\packageIndices{bceQTL}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{bceQTL}

Maintainer: \packageMaintainer{bceQTL}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
date()
}
